+++ admin/class-ilove-img-wm-process.php
